requirements.txt
python-docx
indic-nlp-library
pandas
transformers
torch
tqdm
numpy
scipy
git+https://github.com/VarunGumma/IndicTransToolkit.git
einops
pymupdf
requests
beautifulsoup4
peft
bitsandbytes
datasets
accelerate
PyYAML
